Extract example-mocks crate (#4615)
* add the small crates to solana_program::example_mocks

* extract example-mocks crate

* finish extracting crate

* typo
